IoT-Studio-worker-GRPC/requirements.txt
Flask
grpcio
grpcio-tools
paho-mqtt
openpyxl
grpcio-tools
protobuf